connection to backend - in progress

+++ src/tasks_store.js
@@ -5,16 +5,14 @@ import {PORT} from "./config.js";
 const api_call = Api_call_wrapper.api_call;
 const tasks_api_url = `http://localhost:${PORT}/api/tasks/`
 
-class Tasks_store extends Task {
+class Tasks_store {
   static store = {tasks: [], inited: false};
 
-  constructor() {
-    super();
-  }
-
-  async init_store(){
-    Tasks_store.store.tasks = await api_call(tasks_api_url, 'GET')
-    Tasks_store.store.inited = true
+  async init_store() {
+    const tasks_resp = await api_call(tasks_api_url, 'GET')
+    const tasksData = await tasks_resp.json();
+    Tasks_store.store.tasks = tasksData.map(task => new Task(task));
+    Tasks_store.store.inited = true;
   }
 
   async get_all_tasks() {
@@ -46,32 +44,39 @@ class Tasks_store extends Task {
   }
 
   async add_task(task) {
-    const added_task = await api_call(tasks_api_url, 'POST', task)
-    Tasks_store.store.tasks.push(added_task)
+    const added_task_obj = await api_call(tasks_api_url, 'POST', task)
+    const added_task = await added_task_obj.json()
+    console.log('added_task')
+    console.log(added_task)
+    Tasks_store.store.tasks.push(new Task(added_task))
+    console.log('Tasks_store.store.tasks after adding')
+    console.log(Tasks_store.store.tasks)
   }
 
-  delete_task(id) {
-    // todo: implement the same logic as in add_task
-    let success = false
-    this.get_all_tasks().forEach((task, i) => {
-      if (task.id === id) {
-        task.delete()
-        success = true;
-      }
-    })
-    if (!success)
-      return 'Error while deleting task. Task with provided id wasn\'t found in the task store'
-    return `Task with id ${id} successfully marked as deleted`
+  async delete_task(id) {
+    const response = await api_call(tasks_api_url + id, 'DELETE')
+    if (response.ok) {
+      let success = false
+      const all_tasks = await this.get_all_tasks()
+      all_tasks.forEach((task, i) => {
+        if (task.id === id) {
+          task.delete()
+          success = true;
+        }
+      })
+      if (!success)
+        console.log ('Error while deleting task. Task with provided id wasn\'t found in the local task store')
+      console.log ( `Task with id ${id} successfully marked as deleted in the local task store`)
+    }
   }
 
-  get_task(id) {
-    if (this.get_all_tasks().length > 0) {
-      for (let i = 0; i < this.get_all_tasks().length; i++) {
-        if (this.get_all_tasks()[i].id === id)
-          return this.get_all_tasks()[i]
-      }
+  async get_task(id) {
+    const all_tasks = await this.get_all_tasks();
+    const task = all_tasks.find(t => t.id === id);
+    if (task) {
+      return task;
     } else {
-      throw new Error('Error in get_task function! tasks store is empty')
+      throw new Error(`Error in get_task function! Task with id ${id} not found.`);
     }
   }
 }
